moving v1 api resources for general consistency and swagger generation ease
