+++ multiapps-controller-api/src/main/java/org/cloudfoundry/multiapps/controller/api/v1/CsrfTokenApi.java
@@ -1,7 +1,8 @@
-package org.cloudfoundry.multiapps.controller.api;
+package org.cloudfoundry.multiapps.controller.api.v1;
 
 import javax.inject.Inject;
 
+import org.cloudfoundry.multiapps.controller.api.CsrfTokenApiService;
 import org.cloudfoundry.multiapps.controller.api.Constants.Resources;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
